Partially implemented StudentController

+++ src/main/java/info/oldcolony/shopplacementapp/model/cruds/Shop.java
@@ -55,7 +55,7 @@ public class Shop {
         return true;
     }
 
-    public void unenrollStudentWithId(Long id) {
+    public void unenrollStudentWithId(Integer id) {
         idsOfStudentsEnrolled.remove(id);
     }
